requirements.txt
python-binance
typer[all]
python-dotenv
rich
pydantic
